+++ src/IncompleteTestException.php
@@ -1,0 +1,11 @@
+<?php
+declare(strict_types=1);
+
+namespace MyTester;
+
+/**
+ * @internal
+ */
+class IncompleteTestException extends \RuntimeException
+{
+}